+++ src/main/java/mainProgram/repository/JobPartRepository.java
@@ -1,10 +1,12 @@
 package mainProgram.repository;
 
 import java.util.List;
-import mainProgram.table.JobPart;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import mainProgram.table.JobPart;
+
 public interface JobPartRepository extends JpaRepository<JobPart, Long> {
-  List<JobPart> findByJobId(Long jobId);
+  List<JobPart> findByJobId(int jobId);
   // Spring Data auto-provides findAll(), findById(), save(), delete(), etc.
 }
